Move LoadingBackdrop to components/ui
