+++ airbyte-webapp/src/components/ui/LoadingBackdrop/LoadingBackdrop.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "../../../scss/colors";
 
 %cover {
   width: 100%;
